Migrate to the DB schema v3 (#34)
* Archive the v2 version of the core and DB entities.
* Make the Recipe's "favorite" field mandatory.
* Migrate to the "softDeletionTimestamp" field.
* Add the indexes on "softDeletionTimestamp" field.
* Add "EnhancedCoffeeBeans" table.
* Add the logic to regenerate the "EnhancedCoffeeBeans" items when needed.
* Update the DB CRUD logic for "CoffeeBeans" and "Recipes".
* Add the DB migration code.
* Add the latest/earliest recipe info to the "CoffeeBeansCard".
* Add the logic to sort "CoffeeBeans" by more/less recipes.
* Add a secondary sort order for "Recipes" and "CoffeeBeans".
* Add the "files.exclude" for VS Code's File Explorer settings.
* Add the "CHANGELOG.md" file.
* Add a link to the "CHANGELOG.md" on the "/other" page.
* Change the dates' display to show the month name.
* Move the info row to the bottom of "CoffeeBeansCard".
* Make the "delete all data" button actually delete the DB from the browser, rather than only clearing the object stores.
* Improve the folder structure by moving many files around.
* Improve the "database" folder codebase.
* Change the "enable persistent storage" button's copy.
* Rename the "add a toast" dev buttons.
* Fixe the link color on the "/other" page.

+++ src/lib/helpers/dateHelpers.ts
@@ -9,7 +9,7 @@ export function convertToTimeAgo(date: Date): string {
     if (yesterday.toDateString() === date.toDateString()) {
       return "yesterday";
     }
-    return date.toLocaleString(undefined, { dateStyle: "short" });
+    return date.toLocaleString(undefined, { dateStyle: "long" });
   }
   const hoursAgo = Math.floor(seconds / 3600);
   if (hoursAgo > 1) {
